requirements.txt
langchain
fastapi 
pypdf==5.6.1
python-dotenv
langchain_openai
langchain-pinecone
langchain-community
uvicorn
sentence-transformers
rank-bm25
cohere
sentence-transformers
numpy
scikit-learn